Improve hca zip validation (#4160)
* Use education benefits validation for valid ZIP codes for HCA

* Still use original validation for postal code
