+++ src/js/hca/utils/validations.js
@@ -1,5 +1,6 @@
 import _ from 'lodash';
 import { states } from '../../common/utils/options-for-select';
+import { isValidUSZipCode, isValidCanPostalCode } from '../../common/utils/validations';
 
 function validateIfDirty(field, validator) {
   if (field.dirty) {
@@ -139,12 +140,22 @@ function isValidAddressField(field) {
     isNotBlank(field.city.value) &&
     isNotBlank(field.country.value);
 
+  let isValidPostalCode = true;
+
+  if (field.country.value === 'USA') {
+    isValidPostalCode = isValidPostalCode && isValidRequiredField(isValidUSZipCode, field.zipcode);
+  }
+
+  if (field.country.value === 'CAN') {
+    isValidPostalCode = isValidPostalCode && isValidRequiredField(isValidCanPostalCode, field.zipcode);
+  }
+
   // if we have a defined list of values, they will
   // be set as the state and zipcode keys
   if (_.hasIn(states, field.country.value)) {
     return initialOk &&
       isNotBlank(field.state.value) &&
-      isNotBlank(field.zipcode.value);
+      isValidPostalCode;
   }
   // if the entry was non-USA/CAN/MEX, only postal is
   // required, not provinceCode
